// codestage/src/lib/generated-api/models/JwtCandidate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface JwtCandidate
 */
export interface JwtCandidate {
    /**
     * 
     * @type {string}
     * @memberof JwtCandidate
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof JwtCandidate
     */
    name?: string;
    /**
     * 
     * @type {boolean}
     * @memberof JwtCandidate
     */
    isAdmin?: boolean;
    /**
     * 
     * @type {string}
     * @memberof JwtCandidate
     */
    sessionId?: string;
}

/**
 * Check if a given object implements the JwtCandidate interface.
 */
export function instanceOfJwtCandidate(value: object): value is JwtCandidate {
    return true;
}

export function JwtCandidateFromJSON(json: any): JwtCandidate {
    return JwtCandidateFromJSONTyped(json, false);
}

export function JwtCandidateFromJSONTyped(json: any, ignoreDiscriminator: boolean): JwtCandidate {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
        'name': json['name'] == null ? undefined : json['name'],
        'isAdmin': json['isAdmin'] == null ? undefined : json['isAdmin'],
        'sessionId': json['sessionId'] == null ? undefined : json['sessionId'],
    };
}

export function JwtCandidateToJSON(json: any): JwtCandidate {
    return JwtCandidateToJSONTyped(json, false);
}

export function JwtCandidateToJSONTyped(value?: JwtCandidate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
        'name': value['name'],
        'isAdmin': value['isAdmin'],
        'sessionId': value['sessionId'],
    };
}